refactor(pi-conductor): extract objective orchestration service
