+++ packages/pi-conductor/extensions/conductor.ts
@@ -14,7 +14,8 @@ import {
   validatePushPreconditions,
 } from "./git-pr.js";
 import { computeNextActions, incompleteDependenciesForTask } from "./next-actions.js";
-import { deriveProjectKey } from "./project-key.js";
+import { planObjectiveForRepo, refreshObjectiveStatusForRepo } from "./objective-service.js";
+import { getOrCreateRunForRepo, mutateRepoRunSync } from "./repo-run.js";
 import {
   createWorkerSessionRuntime,
   preflightWorkerRunRuntime,
@@ -24,7 +25,6 @@ import {
 import { type SchedulerFairness, type SchedulerPolicyName, selectSchedulerActions } from "./scheduler-selection.js";
 import {
   addConductorArtifact,
-  addObjective,
   addTask,
   addWorker,
   appendConductorEvent,
@@ -32,13 +32,10 @@ import {
   cancelTaskRun,
   completeTaskRun,
   createConductorGate,
-  createObjectiveRecord,
   createTaskRecord,
   createWorkerRecord,
   linkTaskToObjective,
   markConductorGateUsed,
-  mutateRunWithFileLockSync,
-  readRun,
   reconcileRunLeases,
   recordTaskCompletion,
   recordTaskProgress,
@@ -46,7 +43,6 @@ import {
   resolveConductorGate,
   setWorkerPrState,
   startTaskRun,
-  updateObjective,
   updateTask,
 } from "./storage.js";
 import type {
@@ -62,9 +58,7 @@ import type {
   EvidenceBundlePurpose,
   GateRecord,
   GateStatus,
-  ObjectivePlanResult,
   ObjectiveRecord,
-  ObjectiveStatus,
   ReadinessCheck,
   ReadinessPurpose,
   RunAttemptRecord,
@@ -76,6 +70,14 @@ import type {
 } from "./types.js";
 
 export { computeNextActions } from "./next-actions.js";
+export {
+  createObjectiveForRepo,
+  linkTaskToObjectiveForRepo,
+  planObjectiveForRepo,
+  refreshObjectiveStatusForRepo,
+  updateObjectiveForRepo,
+} from "./objective-service.js";
+export { getOrCreateRunForRepo } from "./repo-run.js";
 export type { SchedulerFairness, SchedulerPolicyName } from "./scheduler-selection.js";
 
 function resolveSchedulerPolicy(input: { policy?: SchedulerPolicyName; executeRuns?: boolean }): {
@@ -86,11 +88,6 @@ function resolveSchedulerPolicy(input: { policy?: SchedulerPolicyName; executeRu
     return { policy: input.policy, executeRuns: input.policy === "execute" };
   }
   return { policy: input.executeRuns ? "execute" : "safe", executeRuns: input.executeRuns ?? false };
-}
-
-function mutateRepoRunSync(repoRoot: string, mutator: (run: RunRecord) => RunRecord): RunRecord {
-  const normalizedRoot = resolve(repoRoot);
-  return mutateRunWithFileLockSync(deriveProjectKey(normalizedRoot), normalizedRoot, mutator);
 }
 
 function errorMessage(error: unknown): string {
@@ -819,17 +816,6 @@ export function getNextActionsForRepo(
   return computeNextActions(run, { ...input, reconciledPreview: input.reconcile ?? true });
 }
 
-export function getOrCreateRunForRepo(repoRoot: string): RunRecord {
-  const normalizedRoot = resolve(repoRoot);
-  const projectKey = deriveProjectKey(normalizedRoot);
-  const existing = readRun(projectKey);
-  return existing ?? mutateRepoRunSync(normalizedRoot, (run) => run);
-}
-
-function createObjectiveId(): string {
-  return `objective-${Date.now().toString(36)}-${Math.random().toString(36).slice(2, 8)}`;
-}
-
 function createTaskId(): string {
   return `task-${Date.now().toString(36)}-${Math.random().toString(36).slice(2, 8)}`;
 }
@@ -840,159 +826,6 @@ function createRunId(): string {
 
 function leaseExpiryFromNow(leaseSeconds: number): string {
   return new Date(Date.now() + leaseSeconds * 1000).toISOString();
-}
-
-export function createObjectiveForRepo(repoRoot: string, input: { title: string; prompt: string }): ObjectiveRecord {
-  let objective!: ObjectiveRecord;
-  mutateRepoRunSync(repoRoot, (run) => {
-    objective = createObjectiveRecord({
-      objectiveId: createObjectiveId(),
-      title: input.title,
-      prompt: input.prompt,
-    });
-    return addObjective(run, objective);
-  });
-  return objective;
-}
-
-export function updateObjectiveForRepo(
-  repoRoot: string,
-  input: { objectiveId: string; title?: string; prompt?: string; status?: ObjectiveStatus; summary?: string | null },
-): ObjectiveRecord {
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => updateObjective(run, input));
-  const objective = updatedRun.objectives.find((entry) => entry.objectiveId === input.objectiveId);
-  if (!objective) {
-    throw new Error(`Objective ${input.objectiveId} disappeared during update`);
-  }
-  return objective;
-}
-
-export function refreshObjectiveStatusForRepo(repoRoot: string, objectiveId: string): ObjectiveRecord {
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => {
-    const objective = run.objectives.find((entry) => entry.objectiveId === objectiveId);
-    if (!objective) {
-      throw new Error(`Objective ${objectiveId} not found`);
-    }
-    const tasks = run.tasks.filter((task) => objective.taskIds.includes(task.taskId));
-    const completed = tasks.filter((task) => task.state === "completed").length;
-    const blocked = tasks.filter((task) => ["blocked", "failed", "canceled"].includes(task.state)).length;
-    const needsReview = tasks.filter((task) => task.state === "needs_review").length;
-    const status =
-      tasks.length > 0 && completed === tasks.length
-        ? "completed"
-        : blocked > 0
-          ? "blocked"
-          : needsReview > 0
-            ? "needs_review"
-            : objective.status === "draft"
-              ? "draft"
-              : "active";
-    const summary = `${completed}/${tasks.length} tasks completed; blocked=${blocked}; needs_review=${needsReview}`;
-    return appendConductorEvent(
-      {
-        ...run,
-        objectives: run.objectives.map((entry) =>
-          entry.objectiveId === objectiveId
-            ? { ...entry, status, summary, revision: entry.revision + 1, updatedAt: new Date().toISOString() }
-            : entry,
-        ),
-      },
-      {
-        actor: { type: "system", id: "conductor" },
-        type: "objective.status_refreshed",
-        resourceRefs: { projectKey: run.projectKey, objectiveId },
-        payload: { status, summary },
-      },
-    );
-  });
-  const updated = updatedRun.objectives.find((entry) => entry.objectiveId === objectiveId);
-  if (!updated) {
-    throw new Error(`Objective ${objectiveId} disappeared during status refresh`);
-  }
-  return updated;
-}
-
-export function linkTaskToObjectiveForRepo(repoRoot: string, objectiveId: string, taskId: string): ObjectiveRecord {
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => linkTaskToObjective(run, objectiveId, taskId));
-  const objective = updatedRun.objectives.find((entry) => entry.objectiveId === objectiveId);
-  if (!objective) {
-    throw new Error(`Objective ${objectiveId} disappeared during link`);
-  }
-  return objective;
-}
-
-function validateObjectivePlanTasks(tasks: Array<{ title: string; prompt: string; dependsOn?: string[] }>): void {
-  const titles = new Set<string>();
-  for (const task of tasks) {
-    const title = task.title.trim();
-    if (titles.has(title)) {
-      throw new Error(`Duplicate task title '${title}' in objective plan`);
-    }
-    titles.add(title);
-    if (task.prompt.trim().length < 10) {
-      throw new Error(`Vague prompt for task '${title}' in objective plan`);
-    }
-  }
-  for (const task of tasks) {
-    for (const dependency of task.dependsOn ?? []) {
-      if (!titles.has(dependency)) {
-        throw new Error(`Unresolved dependency '${dependency}' for task '${task.title}'`);
-      }
-      if (dependency === task.title) {
-        throw new Error(`Task '${task.title}' cannot depend on itself`);
-      }
-    }
-  }
-}
-
-export function planObjectiveForRepo(
-  repoRoot: string,
-  input: {
-    objectiveId: string;
-    tasks: Array<{ title: string; prompt: string; dependsOn?: string[] }>;
-    rationale?: string;
-  },
-): ObjectivePlanResult {
-  if (input.tasks.length === 0) {
-    throw new Error("Objective plan must include at least one task");
-  }
-  validateObjectivePlanTasks(input.tasks);
-  const createdTasks: TaskRecord[] = [];
-  const updatedRun = mutateRepoRunSync(repoRoot, (run) => {
-    const objective = run.objectives.find((entry) => entry.objectiveId === input.objectiveId);
-    if (!objective) {
-      throw new Error(`Objective ${input.objectiveId} not found`);
-    }
-    let nextRun = run;
-    for (const taskInput of input.tasks) {
-      const dependencyText = taskInput.dependsOn?.length ? `\n\nDepends on: ${taskInput.dependsOn.join(", ")}` : "";
-      const task = createTaskRecord({
-        taskId: createTaskId(),
-        title: taskInput.title,
-        prompt: `${taskInput.prompt}${dependencyText}`,
-        objectiveId: input.objectiveId,
-        dependsOnTaskIds: taskInput.dependsOn
-          ?.map(
-            (dependency) =>
-              createdTasks.find((task) => task.title === dependency || task.taskId === dependency)?.taskId,
-          )
-          .filter((taskId): taskId is string => Boolean(taskId)),
-      });
-      nextRun = linkTaskToObjective(addTask(nextRun, task), input.objectiveId, task.taskId);
-      createdTasks.push(task);
-    }
-    return appendConductorEvent(nextRun, {
-      actor: { type: "parent_agent", id: "conductor" },
-      type: "objective.planned",
-      resourceRefs: { projectKey: nextRun.projectKey, objectiveId: input.objectiveId },
-      payload: { taskIds: createdTasks.map((task) => task.taskId), rationale: input.rationale ?? null },
-    });
-  });
-  const updatedObjective = updatedRun.objectives.find((entry) => entry.objectiveId === input.objectiveId);
-  if (!updatedObjective) {
-    throw new Error(`Objective ${input.objectiveId} disappeared during planning`);
-  }
-  return { objective: updatedObjective, tasks: createdTasks };
 }
 
 export function createTaskForRepo(
